Catch errors around document rendering
