--- app/components/ErrorBoundary/index.js
@@ -1,3 +1,0 @@
-// @flow
-import ErrorBoundary from './ErrorBoundary';
-export default ErrorBoundary;
